Tests nachgezogen

+++ src/test/java/ch/zhaw/hs/thesisvalidator/AxiomMapperTest.java
@@ -4,6 +4,7 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
+import java.math.BigInteger;
 import java.util.Map;
 
 import org.jmock.Expectations;
@@ -91,68 +92,68 @@ public class AxiomMapperTest {
 	@Test
 	public void shouldSatisfyNeutralElementForStandardModulo2() {
 		AxiomMapper mapper = new AxiomMapper();
-		assertTrue(mapper.neutral(2, 6, 0));
+		assertTrue(mapper.neutral(2, BigInteger.valueOf(6), 0));
 	}
 	
 	@Test
 	public void shouldNotSatisfyNotNeutralElementForStandardModulo2() {
 		AxiomMapper mapper = new AxiomMapper();
-		assertFalse(mapper.neutral(2, 6, 1));
+		assertFalse(mapper.neutral(2, BigInteger.valueOf(6), 1));
 	}
 	
 	@Test
 	public void shouldNotSatisfyAnyNeutralInAllZeroMappingForMod2() {
 		AxiomMapper mapper = new AxiomMapper();
-		assertFalse(mapper.neutral(2, 0, 0));
-		assertFalse(mapper.neutral(2, 0, 1));
+		assertFalse(mapper.neutral(2, BigInteger.valueOf(0), 0));
+		assertFalse(mapper.neutral(2, BigInteger.valueOf(0), 1));
 	}
 
 	@Test
 	public void shouldSatisfyInverseAxiomForStandardModulo2() {
 		AxiomMapper mapper = new AxiomMapper();
 		// ich nehme an, die fuenfte permutation ist die intuitive additionstabelle der restklasse 2
-		assertTrue(mapper.inverse(2, 6, 0));
+		assertTrue(mapper.inverse(2, BigInteger.valueOf(6), 0));
 	}
 	
 	@Test
 	public void shouldSatisfyAssociativityAxiomForStandardModulo2() {
 		AxiomMapper mapper = new AxiomMapper();
 		// ich nehme an, die fuenfte permutation ist die intuitive additionstabelle der restklasse 2
-		assertTrue(mapper.associative(2, 6));
+		assertTrue(mapper.associative(2, BigInteger.valueOf(6)));
 	}
 	
 	@Test
 	public void shouldSatisfyNeutralElementForStandardModulo3() {
 		AxiomMapper mapper = new AxiomMapper();
-		assertTrue(mapper.neutral(3, 4069, 0));
+		assertTrue(mapper.neutral(3, BigInteger.valueOf(4069), 0));
 	}
 
 	@Test
 	public void shouldNotSatisfyNotNeutralElementForStandardModulo3() {
 		AxiomMapper mapper = new AxiomMapper();
-		assertFalse(mapper.neutral(3, 4069, 1));
-		assertFalse(mapper.neutral(3, 4069, 2));
+		assertFalse(mapper.neutral(3, BigInteger.valueOf(4069), 1));
+		assertFalse(mapper.neutral(3, BigInteger.valueOf(4069), 2));
 	}
 	
 	@Test
 	public void shouldNotSatisfyAnyNeutralInAllZeroMappingForMod3() {
 		AxiomMapper mapper = new AxiomMapper();
-		assertFalse(mapper.neutral(3, 0, 0));
-		assertFalse(mapper.neutral(3, 0, 1));
-		assertFalse(mapper.neutral(3, 0, 2));
+		assertFalse(mapper.neutral(3, BigInteger.valueOf(0), 0));
+		assertFalse(mapper.neutral(3, BigInteger.valueOf(0), 1));
+		assertFalse(mapper.neutral(3, BigInteger.valueOf(0), 2));
 	}
 
 
 	@Test
 	public void shouldSatisfyInverseAxiomForStandardModulo3() {
 		AxiomMapper mapper = new AxiomMapper();
-		assertTrue(mapper.inverse(3, 4069, 0));
+		assertTrue(mapper.inverse(3, BigInteger.valueOf(4069), 0));
 	}
 	
 	@Test
 	public void shouldSatisfyAssociativityAxiomForStandardModulo3() {
 		AxiomMapper mapper = new AxiomMapper();
-		assertTrue(mapper.associative(3, 4069));
+		assertTrue(mapper.associative(3, BigInteger.valueOf(4069)));
 	}
 
 	@Test
